@@ -1,5 +1,7 @@
 part of '../keycloak_wrapper.dart';
 
 const _appAuth = FlutterAppAuth();
-const _refreshTokenKey = 'keycloak-refresh-token';
+const _hasRunBeforeKey = '$_packageName:hasRunBefore';
+const _packageName = 'keycloak_wrapper';
+const _refreshTokenKey = '$_packageName:refreshToken';
 const _secureStorage = FlutterSecureStorage();
